Add simple frontend
This adds a simple frontend application on top of the controller api that allows
for a visual representation of the status of fiaas-enabled namespaces.

The frontend is uses flask-bootstrap that provides bootstrap style templates and jquery.

Split up different flask blueprints.

Embed package data and avoid safe zip due to loading templates and other assets.

Implement first version of deploy and status views.