Extend status and track deployments
This will set a flag indicating a deployment is being performed in a given namespace.
A thread is started that will using a short interval check status for the deployment
and indicate when it is completed. This will allow to communicate the current deployment
status without having to check status in all namespaces simultaneously.