Explicitly set docker image tag in helm chart
This updates the helm/fiaas-skipper/values.yaml file in place with
version reference for the image.tag before packaging the helm chart.

This will ensure that the helm chart that will be published contains a
reference to the docker image tag with the version reference instead of
relying on latest.