Tag releases and publish on release
This introduces automatic tagging with semver on merges to master which
will generate releases based on pre-existings tags. The utility used
will automatically increment the patch version of the last tag: v0.0.1
to v0.0.2 or v0.1.0 to v.0.1.1.
The semver tag will be used to augment the helm chart version as well to
match the tag (excluding the v prefix) for better tracability.
When a tag has been pushed a release is created and docker image and
helm chart published.