Use $ to get root scope when rendering configmap
This fixes a nil reference issue with template rendering when the
`fiaasDeployDaemonConfigMaps` value is used to configure the content of
fiaas-deploy-daemon configmaps.

When changing scope in a template (with range in this case) it is no
longer possible to refer directly to e.g. `.Values.foo`. To refer to a value in
the root scope, we have to use `$.Values.foo`.